+++ crates/authly-client/src/background_worker.rs
@@ -3,27 +3,38 @@ use std::{sync::Arc, time::Duration};
 use authly_common::proto::service::{self as proto};
 use tonic::Streaming;
 
-use crate::{access_control, connection::make_connection, error, ClientState, Error};
+use crate::{
+    access_control,
+    connection::{make_connection, ConnectionParams},
+    error, ClientState, Error,
+};
 
 pub async fn spawn_background_worker(
     state: Arc<ClientState>,
+    reconfigured_tx: tokio::sync::watch::Sender<Arc<ConnectionParams>>,
     closed_rx: tokio::sync::watch::Receiver<()>,
 ) -> Result<(), Error> {
     let msg_stream = init_message_stream(&state).await?;
-    tokio::spawn(background_worker(state, closed_rx, msg_stream));
+    tokio::spawn(background_worker(
+        state,
+        reconfigured_tx,
+        closed_rx,
+        msg_stream,
+    ));
 
     Ok(())
 }
 
 async fn background_worker(
     state: Arc<ClientState>,
+    reconfigured_tx: tokio::sync::watch::Sender<Arc<ConnectionParams>>,
     mut closed_rx: tokio::sync::watch::Receiver<()>,
     mut msg_stream: Streaming<proto::ServiceMessage>,
 ) {
     loop {
         tokio::select! {
             msg_result = msg_stream.message() => {
-                handle_message_result(&state, msg_result, &mut msg_stream).await;
+                handle_message_result(&state, msg_result, &mut msg_stream, &reconfigured_tx).await;
             }
             _ = closed_rx.changed() => {
                 tracing::info!("Authly channel closed");
@@ -37,18 +48,19 @@ async fn handle_message_result(
     state: &ClientState,
     msg_result: Result<Option<proto::ServiceMessage>, tonic::Status>,
     msg_stream: &mut Streaming<proto::ServiceMessage>,
+    reconfigured_tx: &tokio::sync::watch::Sender<Arc<ConnectionParams>>,
 ) {
     match msg_result {
         Ok(Some(msg)) => {
             if let Some(kind) = msg.service_message_kind {
-                handle_message_kind(state, kind, msg_stream).await;
+                handle_message_kind(state, kind, msg_stream, reconfigured_tx).await;
             }
         }
         Ok(None) => {
-            reconfigure_loop(state, msg_stream).await;
+            reconfigure_loop(state, msg_stream, reconfigured_tx).await;
         }
         Err(_error) => {
-            reconfigure_loop(state, msg_stream).await;
+            reconfigure_loop(state, msg_stream, reconfigured_tx).await;
         }
     }
 }
@@ -57,12 +69,13 @@ async fn handle_message_kind(
     state: &ClientState,
     msg_kind: proto::service_message::ServiceMessageKind,
     msg_stream: &mut Streaming<proto::ServiceMessage>,
+    reconfigured_tx: &tokio::sync::watch::Sender<Arc<ConnectionParams>>,
 ) {
     tracing::info!(?msg_kind, "Received Authly message");
 
     match msg_kind {
         proto::service_message::ServiceMessageKind::ReloadCa(_) => {
-            reconfigure_loop(state, msg_stream).await;
+            reconfigure_loop(state, msg_stream, reconfigured_tx).await;
         }
         proto::service_message::ServiceMessageKind::ReloadCache(_) => {
             reload_local_cache(state).await;
@@ -79,9 +92,13 @@ async fn handle_message_kind(
     }
 }
 
-async fn reconfigure_loop(state: &ClientState, msg_stream: &mut Streaming<proto::ServiceMessage>) {
+async fn reconfigure_loop(
+    state: &ClientState,
+    msg_stream: &mut Streaming<proto::ServiceMessage>,
+    reconfigured_tx: &tokio::sync::watch::Sender<Arc<ConnectionParams>>,
+) {
     loop {
-        match try_reconfigure(state, msg_stream).await {
+        match try_reconfigure(state, msg_stream, reconfigured_tx).await {
             Ok(()) => return,
             Err(err) => {
                 tracing::error!(?err, "background reconfigure error");
@@ -95,14 +112,19 @@ async fn reconfigure_loop(state: &ClientState, msg_stream: &mut Streaming<proto:
 async fn try_reconfigure(
     state: &ClientState,
     msg_stream: &mut Streaming<proto::ServiceMessage>,
+    reconfigured_tx: &tokio::sync::watch::Sender<Arc<ConnectionParams>>,
 ) -> Result<(), Error> {
-    let connection_params = state.reconfigure.new_connection_params().await?;
-    let connection = Arc::new(make_connection(connection_params).await?);
+    let params = state.reconfigure.new_connection_params().await?;
+    let connection = Arc::new(make_connection(params.clone()).await?);
 
     state.conn.store(connection.clone());
 
     *msg_stream = init_message_stream(state).await?;
     reload_local_cache(state).await;
+
+    if let Err(err) = reconfigured_tx.send(params) {
+        tracing::error!(?err, "Could not publish reconfigured connection params");
+    }
 
     Ok(())
 }